+++ lib/zego_express_engine.dart
@@ -15,6 +15,7 @@ export 'src/zego_express_api_preprocess.dart';
 export 'src/zego_express_api_im.dart';
 export 'src/zego_express_api_record.dart';
 export 'src/zego_express_api_custom_video_io.dart';
+export 'src/zego_express_api_utilities.dart';
 
 export 'src/zego_express_texture_render_utils.dart';
 export 'src/zego_express_platform_view_utils.dart';